Since DisableWhileTyping is not working, disable Tapping
Touching the touchpad with my palm when typing is a real
issue. DisableWhileTyping isn't doing anything for me, and I can't
even find a mention of palm detection in `xinput --watch-props`. This
seems the only solution; accidentally clicking is annoying, but
accidentally moving the cursor slightly is no problem at all. I don't
really use tapping all that much anyways, not when the touchpad click
works well enough, and it does.