don't boot into X. GPU driver issues or smth in new system?
Just boot to the standard, virtual tty. Maybe user can install gui stuff separately and
open that with xinit or smth if they (I) ever feel the need. But this is a server, so an
actual desktop environment shouldn't be necessary. Can always do X over SSH if there's
truly a need.