on certbot deploy, reload httpd to register updated certs
Noticed today that the TLS certs seemed to have expired when I visited
https://jo.zone. The cert files in /etc/letsencrypt had most recently been updated a month
before firefox said the certs had expired though, sometime in april. Investigation proved
that, indeed, the certs had been updated 1 month befor expire, and apache had simply not
picked up on it. The reason I haven't had this problem before is that, ironically, the
server's uptime has just been "too" good recently -- 45 days. That's not even that much,
of course, but it was enough to miss the cert updates a bit less than a month ago.

These changes should hopefully reload apache on successful cert update. As I understand
it, sending the SIGHUP signal to Apache HTTPD does exactly that. I guess we'll see in a
couple of months...